rename hook from useDAG to useDagStore
