+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useDecisionTree } from './useDecisionTree/useDecisionTree';
-export { useDAG } from './useDAG/useDAG';
+export { useDagStore } from 'hooks/useDagStore/useDagStore';
 export { useFetchConfig } from './useFetchConfig/useFetchConfig';
 export { useTreeDirection } from './useTreeDirection/useTreeDirection';
